+++ schema/app/models/autocomplete.py
@@ -1,4 +1,4 @@
-from sqlalchemy import BigInteger, Column, Index, String
+from sqlalchemy import Column, Index, Integer, String
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -7,7 +7,7 @@ Base = declarative_base()
 class City(Base):
     __tablename__ = "city"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String)
 
     __table_args__ = (Index("ix_city_name", "name"),)
@@ -16,7 +16,7 @@ class City(Base):
 class IndustrySector(Base):
     __tablename__ = "industry_sector"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String)
 
     __table_args__ = (Index("ix_industry_sector_name", "name"),)
@@ -25,7 +25,7 @@ class IndustrySector(Base):
 class LegalForm(Base):
     __tablename__ = "legal_form"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String)
 
     __table_args__ = (Index("ix_legal_form_name", "name"),)
@@ -34,7 +34,7 @@ class LegalForm(Base):
 class Region(Base):
     __tablename__ = "region"
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String)
 
     __table_args__ = (Index("ix_region_name", "name"),)
